python/infraRocket/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
ldap3
itsdangerous
passlib
requests
